🎨 Palette: Enhance task deletion with confirmation dialog and tooltip
- Added a 'Delete Task' tooltip to the task card's delete button for better accessibility.
- Implemented a confirmation AlertDialog in the task list to prevent accidental deletions.
- Included a `context.mounted` check after the dialog to follow Flutter best practices.
- Fixed a numerical literal type mismatch in dashboard metrics to resolve a build error.
- Documented learnings in `.Jules/palette.md`.

+++ lib/features/dashboard/domain/dashboard_metrics.dart
@@ -52,7 +52,7 @@ class DashboardMetrics {
 
     final streakStats = _calculateStreaks(logs);
     final total = completed + skipped + missed;
-    final productivity = total == 0 ? 0 : (completed / total) * 100;
+    final productivity = total == 0 ? 0.0 : (completed / total) * 100;
 
     return DashboardMetrics(
       completed: completed,
@@ -61,7 +61,7 @@ class DashboardMetrics {
       categoryPerformance: categoryPerformance,
       currentStreak: streakStats.$1,
       longestStreak: streakStats.$2,
-      productivityScore: productivity,
+      productivityScore: productivity.toDouble(),
       monthlySummary:
           'Completed $completed tasks this month, skipped $skipped and missed $missed.',
     );
